+++ src/lib/webview/controller/startwork/startWorkWebviewController.ts
@@ -119,7 +119,7 @@ export class StartWorkWebviewController implements WebviewController<StartWorkIs
                 ...this.initData!,
                 repoData,
                 ...this.api.getStartWorkConfig(),
-                isRovoDevEnabled: Container.isRovoDevActive(),
+                isRovoDevEnabled: Container.isRovoDevActive,
             });
         } catch (e) {
             this.logger.error(e, 'Error updating start work page');